+++ app/src/main/java/com/costular/atomtasks/ui/home/AppNavigator.kt
@@ -1,13 +1,11 @@
 package com.costular.atomtasks.ui.home
 
 import androidx.navigation.NavController
-import com.atomtasks.feature.detail.destinations.TaskDetailScreenDestination
-import com.costular.atomtasks.agenda.destinations.TasksActionsBottomSheetDestination
 import com.costular.atomtasks.agenda.ui.AgendaNavigator
-import com.costular.atomtasks.createtask.destinations.CreateTaskScreenDestination
-import com.costular.atomtasks.feature.edittask.destinations.EditTaskScreenDestination
 import com.costular.atomtasks.settings.SettingsNavigator
-import com.costular.atomtasks.settings.destinations.ThemeSelectorScreenDestination
+import com.ramcosta.composedestinations.generated.agenda.destinations.TasksActionsBottomSheetDestination
+import com.ramcosta.composedestinations.generated.detail.destinations.TaskDetailScreenDestination
+import com.ramcosta.composedestinations.generated.settings.destinations.ThemeSelectorScreenDestination
 import com.ramcosta.composedestinations.utils.toDestinationsNavigator
 import java.time.LocalDate
 
@@ -17,14 +15,6 @@ class AppNavigator(
 
     private val destinationsNavigator by lazy {
         navController.toDestinationsNavigator()
-    }
-
-    override fun navigateToCreateTask(date: String) {
-        destinationsNavigator.navigate(CreateTaskScreenDestination(date))
-    }
-
-    override fun navigateToEditTask(taskId: Long) {
-        destinationsNavigator.navigate(EditTaskScreenDestination(taskId))
     }
 
     override fun navigateToDetailScreenForCreateTask(date: String) {