+++ build-logic/convention/src/main/kotlin/AndroidFeatureConventionPlugin.kt
@@ -14,7 +14,6 @@ class AndroidFeatureConventionPlugin : Plugin<Project> {
             pluginManager.apply {
                 apply("com.android.library")
                 apply("org.jetbrains.kotlin.android")
-                apply("org.jetbrains.kotlin.kapt")
             }
             extensions.configure<LibraryExtension>() {
                 defaultConfig {
@@ -39,13 +38,10 @@ class AndroidFeatureConventionPlugin : Plugin<Project> {
                 add("implementation", libs.findLibrary("compose.material.icons").get())
                 add("implementation", libs.findLibrary("compose.ui").get())
                 add("implementation", libs.findLibrary("compose.ui.tooling").get())
-                add("implementation", libs.findLibrary("accompanist.systemui").get())
                 add("implementation", libs.findLibrary("viewmodel").get())
                 add("implementation", libs.findLibrary("hilt.navigation.compose").get())
-                add("kapt", libs.findLibrary("hilt.compiler").get())
-                add("implementation", libs.findLibrary("hilt").get())
-                add("kapt", libs.findLibrary("hilt.compiler").get())
-                add("implementation", libs.findLibrary("compose.destinations").get())
+                add("implementation", libs.findLibrary("compose.destinations.core").get())
+                add("implementation", libs.findLibrary("compose.destinations.bottomsheet").get())
 
                 add("testImplementation", project(":core:testing"))
                 add("androidTestImplementation", project(":core:testing"))