+++ build-logic/convention/src/main/kotlin/com/costular/atomtasks/Versioning.kt
@@ -1,6 +1,6 @@
 package com.costular.atomtasks
 
 object Versioning {
-    const val VersionCode = 23
-    const val VersionName = "2.6.2"
+    const val VersionCode = 24
+    const val VersionName = "2.6.3"
 }