+++ core/designsystem/src/main/java/com/costular/designsystem/components/WeekCalendar.kt
@@ -1,13 +1,12 @@
 package com.costular.designsystem.components
 
-import androidx.compose.foundation.Canvas
+import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.MaterialTheme
@@ -19,7 +18,6 @@ import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.costular.atomtasks.core.ui.date.Day
@@ -84,8 +82,18 @@ private fun CalendarDay(
         CardDefaults.outlinedCardColors()
     }
 
+    val border = if (isToday) {
+        BorderStroke(
+            width = 1.dp,
+            color = MaterialTheme.colorScheme.primary
+        )
+    } else {
+        CardDefaults.outlinedCardBorder()
+    }
+
     OutlinedCard(
         colors = cardColors,
+        border = border,
         modifier = modifier
             .height(70.dp)
             .clip(RoundedCornerShape(12.dp))
@@ -104,29 +112,8 @@ private fun CalendarDay(
                 text = day.toString(),
                 style = MaterialTheme.typography.titleMedium
             )
-
-            if (isToday) {
-                IndicatorToday(
-                    color = MaterialTheme.colorScheme.onPrimaryContainer.copy(alpha = 0.60f)
-                )
-            }
         }
     }
-}
-
-@Composable
-fun IndicatorToday(
-    modifier: Modifier = Modifier,
-    color: Color,
-) {
-    Canvas(
-        modifier = modifier.size(4.dp),
-        onDraw = {
-            drawCircle(
-                color = color,
-            )
-        }
-    )
 }
 
 @Preview