+++ core/ui/tasks/src/main/java/com/costular/atomtasks/core/ui/tasks/TaskList.kt
@@ -1,30 +1,53 @@
 package com.costular.atomtasks.core.ui.tasks
 
-import androidx.compose.foundation.ExperimentalFoundationApi
+import androidx.compose.animation.core.FastOutSlowInEasing
+import androidx.compose.animation.core.animateFloatAsState
+import androidx.compose.animation.core.tween
+import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.LazyItemScope
 import androidx.compose.foundation.lazy.items
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Delete
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.SwipeToDismissBox
+import androidx.compose.material3.SwipeToDismissBoxState
+import androidx.compose.material3.SwipeToDismissBoxValue
 import androidx.compose.material3.Text
+import androidx.compose.material3.rememberSwipeToDismissBoxState
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.derivedStateOf
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.scale
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.platform.LocalView
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import androidx.core.view.HapticFeedbackConstantsCompat
 import com.costular.atomtasks.core.ui.R
 import com.costular.atomtasks.core.ui.utils.VariantsPreview
 import com.costular.atomtasks.tasks.model.Reminder
 import com.costular.atomtasks.tasks.model.Task
 import com.costular.designsystem.theme.AppTheme
 import com.costular.designsystem.theme.AtomTheme
+import kotlinx.coroutines.delay
 import java.time.LocalDate
 import java.time.LocalTime
 import org.burnoutcrew.reorderable.ReorderableItem
@@ -33,11 +56,14 @@ import org.burnoutcrew.reorderable.detectReorderAfterLongPress
 import org.burnoutcrew.reorderable.rememberReorderableLazyListState
 import org.burnoutcrew.reorderable.reorderable
 
-@OptIn(ExperimentalFoundationApi::class)
+private const val MillisecondsToResetSwipeBoxState = 1000L
+
 @Composable
 fun TaskList(
     tasks: List<Task>,
     onClick: (Task) -> Unit,
+    onClickMore: (Task) -> Unit,
+    onDeleteTask: (Task) -> Unit,
     onMarkTask: (taskId: Long, isDone: Boolean) -> Unit,
     state: ReorderableLazyListState,
     modifier: Modifier = Modifier,
@@ -55,22 +81,112 @@ fun TaskList(
             verticalArrangement = Arrangement.spacedBy(8.dp),
         ) {
             items(tasks, { it.id }) { task ->
-                ReorderableItem(state, key = task.id) { isDragging ->
-                    TaskCard(
-                        title = task.name,
-                        onMark = { onMarkTask(task.id, !task.isDone) },
-                        onOpen = { onClick(task) },
-                        reminder = task.reminder,
-                        isFinished = task.isDone,
-                        recurrenceType = task.recurrenceType,
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .animateItemPlacement(),
-                        isBeingDragged = isDragging,
-                    )
-                }
+                TaskItem(
+                    state = state,
+                    task = task,
+                    onDeleteTask = onDeleteTask,
+                    onMarkTask = onMarkTask,
+                    onClick = onClick,
+                    onClickMore = onClickMore
+                )
             }
         }
+    }
+}
+
+@Composable
+private fun LazyItemScope.TaskItem(
+    state: ReorderableLazyListState,
+    task: Task,
+    onDeleteTask: (Task) -> Unit,
+    onMarkTask: (taskId: Long, isDone: Boolean) -> Unit,
+    onClick: (Task) -> Unit,
+    onClickMore: (Task) -> Unit
+) {
+    val dismissState = rememberSwipeToDismissBoxState()
+
+    SwipeToDismissBox(
+        state = dismissState,
+        enableDismissFromEndToStart = true,
+        enableDismissFromStartToEnd = false,
+        backgroundContent = {
+            TaskRemoveBackground(dismissState)
+        },
+    ) {
+        when (dismissState.currentValue) {
+            SwipeToDismissBoxValue.EndToStart -> onDeleteTask(task)
+            SwipeToDismissBoxValue.StartToEnd -> Unit
+            SwipeToDismissBoxValue.Settled -> Unit
+        }
+
+        ReorderableItem(state, key = task.id) { isDragging ->
+            TaskCard(
+                title = task.name,
+                onMark = { onMarkTask(task.id, !task.isDone) },
+                onClick = { onClick(task) },
+                reminder = task.reminder,
+                isFinished = task.isDone,
+                recurrenceType = task.recurrenceType,
+                modifier = Modifier.animateItem(fadeInSpec = null, fadeOutSpec = null),
+                isBeingDragged = isDragging,
+                onClickMore = { onClickMore(task) },
+            )
+        }
+    }
+}
+
+@Composable
+private fun TaskRemoveBackground(
+    state: SwipeToDismissBoxState,
+) {
+    val view = LocalView.current
+    val scale by
+    animateFloatAsState(
+        targetValue = if (state.targetValue == SwipeToDismissBoxValue.Settled) 0.75f else 1f,
+        animationSpec = tween(
+            durationMillis = 200,
+            easing = FastOutSlowInEasing,
+        )
+    )
+
+    val swipeTargetValue by remember {
+        derivedStateOf { state.targetValue }
+    }
+
+    if (state.currentValue != SwipeToDismissBoxValue.Settled) {
+        LaunchedEffect(Unit) {
+            delay(MillisecondsToResetSwipeBoxState)
+            state.reset()
+        }
+    }
+
+    LaunchedEffect(swipeTargetValue) {
+        if (state.targetValue == SwipeToDismissBoxValue.EndToStart) {
+            view.performHapticFeedback(HapticFeedbackConstantsCompat.GESTURE_START)
+        }
+    }
+
+    val (backgroundColor, contentColor) = when (state.dismissDirection) {
+        SwipeToDismissBoxValue.Settled -> Color.Transparent to Color.Transparent
+        SwipeToDismissBoxValue.StartToEnd -> Color.Transparent to Color.Transparent
+        SwipeToDismissBoxValue.EndToStart -> {
+            MaterialTheme.colorScheme.error to MaterialTheme.colorScheme.onError
+        }
+    }
+
+    Box(
+        modifier = Modifier
+            .fillMaxSize()
+            .background(backgroundColor, CardDefaults.elevatedShape)
+            .padding(AppTheme.dimens.contentMargin),
+        contentAlignment = Alignment.CenterEnd,
+    ) {
+        Icon(
+            imageVector = Icons.Default.Delete,
+            contentDescription = null,
+            tint = contentColor,
+            modifier = Modifier.scale(scale),
+        )
     }
 }
 
@@ -166,6 +282,8 @@ private fun TaskListPreview() {
             ),
             onClick = {},
             onMarkTask = { _, _ -> },
+            onClickMore = {},
+            onDeleteTask = {},
         )
     }
 }