+++ data/src/test/java/com/costular/atomtasks/data/settings/SetThemeUseCaseTest.kt
@@ -2,14 +2,11 @@ package com.costular.atomtasks.data.settings
 
 import io.mockk.coVerify
 import io.mockk.mockk
-import kotlinx.coroutines.test.TestCoroutineScope
-import kotlinx.coroutines.test.runBlockingTest
+import kotlinx.coroutines.test.runTest
 import org.junit.Before
 import org.junit.Test
 
 class SetThemeUseCaseTest {
-    private val testScope = TestCoroutineScope()
-
     private val repository: SettingsRepository = mockk(relaxed = true)
 
     lateinit var sut: SetThemeUseCase
@@ -20,12 +17,11 @@ class SetThemeUseCaseTest {
     }
 
     @Test
-    fun `should store theme in preferences when change theme`() =
-        testScope.runBlockingTest {
-            val theme = Theme.Dark
+    fun `should store theme in preferences when change theme`() = runTest {
+        val theme = Theme.Dark
 
-            sut.executeSync(SetThemeUseCase.Params(theme))
+        sut.executeSync(SetThemeUseCase.Params(theme))
 
-            coVerify { repository.setTheme(theme) }
-        }
+        coVerify { repository.setTheme(theme) }
+    }
 }