+++ feature/agenda/src/main/java/com/costular/atomtasks/agenda/ui/AgendaGraph.kt
@@ -1,0 +1,7 @@
+package com.costular.atomtasks.agenda.ui
+
+import com.ramcosta.composedestinations.annotation.ExternalModuleGraph
+import com.ramcosta.composedestinations.annotation.NavGraph
+
+@NavGraph<ExternalModuleGraph>
+internal annotation class AgendaGraph