+++ feature/agenda/src/main/java/com/costular/atomtasks/agenda/ui/AgendaHeader.kt
@@ -1,9 +1,5 @@
 package com.costular.atomtasks.agenda.ui
 
-import androidx.compose.animation.AnimatedContent
-import androidx.compose.animation.core.FastOutSlowInEasing
-import androidx.compose.animation.core.animateFloatAsState
-import androidx.compose.animation.core.tween
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -11,20 +7,16 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.ExpandMore
-import androidx.compose.material.icons.outlined.Today
+import androidx.compose.material.icons.outlined.CalendarMonth
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.rotate
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
@@ -33,7 +25,6 @@ import com.costular.atomtasks.core.ui.R
 import com.costular.atomtasks.core.ui.date.Day
 import com.costular.atomtasks.core.ui.date.asDay
 import com.costular.atomtasks.core.ui.utils.DateUtils
-import com.costular.designsystem.components.DatePicker
 import com.costular.designsystem.components.ScreenHeader
 import com.costular.designsystem.components.WeekCalendar
 import com.costular.designsystem.theme.AppTheme
@@ -53,9 +44,7 @@ internal fun AgendaHeader(
     selectedDay: Day,
     onSelectDate: (LocalDate) -> Unit,
     onSelectToday: () -> Unit,
-    canExpand: Boolean,
-    isExpanded: Boolean,
-    onToggleExpandCollapse: () -> Unit,
+    onClickCalendar: () -> Unit,
 ) {
     val coroutineScope = rememberCoroutineScope()
     val startDate = remember(selectedDay) { selectedDay.date.minusDays(DaysToShow) }
@@ -67,18 +56,8 @@ internal fun AgendaHeader(
         firstVisibleWeekDate = selectedDay.date,
     )
 
-    val shadowElevation = if (isExpanded) {
-        6.dp
-    } else {
-        2.dp
-    }
-
-    Surface(
-        modifier = modifier,
-        shape = RoundedCornerShape(bottomStart = 16.dp, bottomEnd = 16.dp),
-        shadowElevation = shadowElevation,
-    ) {
-        Column {
+    Surface {
+        Column(modifier) {
             Row(
                 modifier = Modifier.fillMaxWidth(),
                 verticalAlignment = Alignment.CenterVertically,
@@ -89,7 +68,12 @@ internal fun AgendaHeader(
                     verticalAlignment = Alignment.CenterVertically,
                     modifier = Modifier
                         .weight(1f)
-                        .clickable(enabled = canExpand, onClick = onToggleExpandCollapse),
+                        .clickable(onClick = {
+                            coroutineScope.launch {
+                                weekCalendarState.animateScrollToWeek(LocalDate.now())
+                            }
+                            onSelectToday()
+                        }),
                 ) {
                     ScreenHeader(
                         text = selectedDayText,
@@ -101,81 +85,33 @@ internal fun AgendaHeader(
                                 start = AppTheme.dimens.spacingLarge,
                             ),
                     )
-
-                    val degrees by animateFloatAsState(
-                        targetValue = if (isExpanded) 180f else 0f,
-                        animationSpec = tween(
-                            durationMillis = 200,
-                            easing = FastOutSlowInEasing,
-                        ),
-                        label = "Header collapsable arrow",
-                    )
-
-                    if (canExpand) {
-                        IconButton(onClick = { onToggleExpandCollapse() }) {
-                            Icon(
-                                imageVector = Icons.Default.ExpandMore,
-                                contentDescription = null,
-                                modifier = Modifier.rotate(degrees),
-                            )
-                        }
-                    }
                 }
 
                 IconButton(
-                    onClick = {
-                        coroutineScope.launch {
-                            weekCalendarState.animateScrollToWeek(LocalDate.now())
-                        }
-                        onSelectToday()
-                    },
+                    onClick = onClickCalendar,
                     modifier = Modifier
                         .padding(end = AppTheme.dimens.spacingLarge)
                         .width(40.dp)
                         .height(40.dp),
                 ) {
                     Icon(
-                        imageVector = Icons.Outlined.Today,
-                        contentDescription = stringResource(R.string.today),
+                        imageVector = Icons.Outlined.CalendarMonth,
+                        contentDescription = stringResource(R.string.home_menu_calendar),
                     )
                 }
             }
 
-            AnimatedContent(
-                targetState = isExpanded && canExpand,
-                label = "Header calendar",
-            ) { isCollapsed ->
-                if (isCollapsed) {
-                    ExpandedCalendar(selectedDay, onSelectDate)
-                } else {
-                    CollapsedCalendar(
-                        selectedDay,
-                        onSelectDate,
-                        weekCalendarState,
-                    )
-                }
-            }
+            WeekCalendar(
+                selectedDay,
+                onSelectDate,
+                weekCalendarState,
+            )
         }
     }
 }
 
 @Composable
-private fun ExpandedCalendar(
-    selectedDay: Day,
-    onSelectDate: (LocalDate) -> Unit,
-) {
-    DatePicker(
-        modifier = Modifier
-            .supportWideScreen(480.dp)
-            .padding(horizontal = AppTheme.dimens.contentMargin)
-            .padding(bottom = AppTheme.dimens.spacingMedium),
-        selectedDay = selectedDay.date,
-        onDateSelected = onSelectDate,
-    )
-}
-
-@Composable
-private fun CollapsedCalendar(
+private fun WeekCalendar(
     selectedDay: Day,
     onSelectDate: (LocalDate) -> Unit,
     weekCalendarState: WeekCalendarState = rememberWeekCalendarState(),
@@ -186,7 +122,7 @@ private fun CollapsedCalendar(
             .padding(
                 start = AppTheme.dimens.spacingLarge,
                 end = AppTheme.dimens.spacingLarge,
-                bottom = AppTheme.dimens.spacingLarge,
+                bottom = AppTheme.dimens.spacingSmall,
             ),
         selectedDay = selectedDay,
         onSelectDay = onSelectDate,
@@ -202,26 +138,8 @@ private fun HeaderCollapsedPreview() {
             selectedDay = LocalDate.now().asDay(),
             onSelectDate = {},
             onSelectToday = {},
-            canExpand = true,
-            isExpanded = false,
-            onToggleExpandCollapse = {},
             modifier = Modifier.fillMaxWidth(),
-        )
-    }
-}
-
-@Preview
-@Composable
-private fun HeaderExpandedPreview() {
-    AtomTheme {
-        AgendaHeader(
-            selectedDay = LocalDate.now().asDay(),
-            onSelectDate = {},
-            onSelectToday = {},
-            canExpand = true,
-            isExpanded = true,
-            onToggleExpandCollapse = {},
-            modifier = Modifier.fillMaxWidth(),
+            onClickCalendar = {},
         )
     }
 }