+++ feature/agenda/src/main/java/com/costular/atomtasks/agenda/ui/AgendaNavigator.kt
@@ -1,14 +1,6 @@
 package com.costular.atomtasks.agenda.ui
 
 interface AgendaNavigator {
-    fun navigateToCreateTask(
-        date: String,
-    )
-
-    fun navigateToEditTask(
-        taskId: Long,
-    )
-
     fun navigateToDetailScreenForCreateTask(
         date: String,
     )