+++ feature/createtask/src/main/java/com/costular/atomtasks/createtask/CreateTaskState.kt
@@ -1,9 +1,0 @@
-package com.costular.atomtasks.createtask
-
-sealed interface CreateTaskState {
-
-    data object Uninitialized : CreateTaskState
-    data object Saving : CreateTaskState
-    data object Success : CreateTaskState
-    data object Failure : CreateTaskState
-}