+++ feature/settings/src/main/java/com/costular/atomtasks/settings/SettingsGraph.kt
@@ -1,0 +1,7 @@
+package com.costular.atomtasks.settings
+
+import com.ramcosta.composedestinations.annotation.ExternalModuleGraph
+import com.ramcosta.composedestinations.annotation.NavGraph
+
+@NavGraph<ExternalModuleGraph>
+internal annotation class SettingsGraph